Added tests to the RS232 and USB projects
I created the test project for the RS232 project, and added a test to
the USB project. I also fixed a bug that was present in the USB layer of
both of them.

I updated the RS232 Arduino project a bit, just to format the log
statements a bit better.