Made a library for mocked objects, fixed bugs
The testing was all messed up. I changed the implementation of the
mocked sreial objects to use strings, but then memory got way out of
hand during testing. I changed how the testing is done for BTRS232, but
the card reader tests still need to be changed. The Mocks library also
needs a bit of cleaning.